Refactor smart glass dialog menu themes
