+++ bandyer-android-design/src/main/res/layout/bandyer_widget_smartglasses_menu_layout.xml
@@ -7,12 +7,8 @@
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/bandyer_smart_glass_menu_recyclerview"
         android:layout_gravity="center"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        android:layout_height="0dp"
-        android:layout_width="0dp" />
+        android:layout_height="match_parent"
+        android:layout_width="match_parent" />
 
     <com.bandyer.sdk_design.smartglass.call.menu.SmartGlassMenuPagerIndicator
         style="?attr/bandyer_pagerIndicatorStyle"
