+++ bandyer-android-design/src/main/res/values/smartglass_menu.xml
@@ -26,9 +26,9 @@
     <style name="BandyerSDKDesign.SmartGlassDialogMenu">
         <item name="android:background">?attr/colorSurface</item>
         <item name="android:paddingBottom">@dimen/bandyer_dimen_space16</item>
-        <item name="bandyer_smartGlassMicrophoneStyle">@style/BandyerSDKDesign.ActionButton.Smartglass.Microphone</item>
-        <item name="bandyer_smartGlassCameraStyle">@style/BandyerSDKDesign.ActionButton.Smartglass.Camera</item>
-        <item name="bandyer_smartGlassParticipantsStyle">@style/BandyerSDKDesign.ActionButton.Smartglass.Participants</item>
+        <item name="bandyer_smartGlassMicrophoneStyle">@style/BandyerSDKDesign.ActionButton.SmartGlass.Microphone</item>
+        <item name="bandyer_smartGlassCameraStyle">@style/BandyerSDKDesign.ActionButton.SmartGlass.Camera</item>
+        <item name="bandyer_smartGlassParticipantsStyle">@style/BandyerSDKDesign.ActionButton.SmartGlass.Participants</item>
         <item name="bandyer_pagerIndicatorStyle">@style/BandyerSDKDesign.PagerIndicatorStyle</item>
     </style>
 </resources>
