+++ demo_app/src/main/res/values-it/strings.xml
@@ -32,5 +32,5 @@
     <string name="bandyer_start_video_call_description">Avvia videochiamata</string>
     <string name="bandyer_start_audio_upgrade_call_description">Avvia videochiamata con solo audio</string>
     <string name="screen_share_picker">Modalità screen share</string>
-    <string name="smartglasses_menu">Smartglass menu</string>
+    <string name="smart_glass_menu">Smartglass menu</string>
 </resources>
